updated content.js and news.js
updated the two mentioned file so that the database acts as the serving point for news cards rather than the api. this will reduce the load on the api and allow us to quickly serve cached news. however still need to update the get requests to no longer ping the api at every instances.